Refactor App.js into multiple modules
I've split the main App.js file into smaller, more manageable modules for contexts, hooks, components (screens and common), constants, and services.

- I created a new directory structure for better organization.
- I moved relevant code segments to their respective new files.
- I updated App.js to import new modules and retain the main app structure.
- I updated import/export statements across all affected files.

Note: Build verification was inconclusive due to a persistent 'react-dom/client' resolution error, which I suspect is an environment or tooling issue. The refactored code structure itself aligns with your requested changes.
